General cleanup and Refactor.
Added translations to gui buttons instead of hard-coded English.
