+++ src/main/java/ninja/genuine/tooltips/client/config/Config.java
@@ -1,4 +1,4 @@
-package ninja.genuine.tooltips;
+package ninja.genuine.tooltips.client.config;
 
 import java.util.Objects;
 
@@ -6,7 +6,6 @@ import net.minecraftforge.common.config.ConfigCategory;
 import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.common.config.Property;
 import net.minecraftforge.common.config.Property.Type;
-import ninja.genuine.tooltips.client.gui.GuiConfigTooltips.ColorEntry;
 
 public class Config {
 
@@ -14,13 +13,6 @@ public class Config {
 	public static final String category_general = "General";
 	public static final String category_appearance = "Appearance";
 	public static final String category_behavior = "Behavior";
-	public static final String name_enable = "Enable Mod";
-	public static final String name_distance = "Maximum Drawing Distance";
-	public static final String name_override_outline = "Should Override Outline";
-	public static final String name_hide_mod = "Hide Mod Name";
-	public static final String name_opacity = "Tooltip Opacity";
-	public static final String name_outline = "Outline Color";
-	public static final String name_background = "Background Color";
 
 	public static Config getInstance() {
 		if (Objects.isNull(instance))
@@ -28,11 +20,11 @@ public class Config {
 		return instance;
 	}
 
-	static void setConfiguration(Configuration config) {
+	public static void setConfiguration(Configuration config) {
 		getInstance().base = config;
 	}
 
-	static void save() {
+	public static void save() {
 		getInstance().base.save();
 	}
 
@@ -40,40 +32,40 @@ public class Config {
 		return getInstance().base.getCategory(name);
 	}
 
+	public static void populate() {
+		getInstance().internalPopulation();
+	}
+
 	private Configuration base;
 
 	private Config() {}
 
-	static void populate() {
-		getInstance().internalPopulation();
-	}
-
 	public boolean isEnabled() {
-		return base.getBoolean(name_enable, category_general, true, "Enable rendering the tooltips.");
+		return base.getBoolean("Enable Mod", category_general, true, "Enable rendering the tooltips.");
 	}
 
 	public int getMaxDistance() {
-		return base.getInt(name_distance, category_behavior, 8, 2, 64, "Set the maximum distance that tooltips should be displayed from.");
+		return base.getInt("Maximum Drawing Distance", category_behavior, 8, 2, 64, "Set the maximum distance that tooltips should be displayed from.");
 	}
 
 	public boolean isOverridingOutline() {
-		return base.getBoolean(name_override_outline, category_behavior, false, "If enabled outline color will be manually set instead of default behavior.");
+		return base.getBoolean("Override Outline Color", category_behavior, false, "If enabled outline color will be manually set instead of default behavior.");
 	}
 
 	public boolean isHidingMod() {
-		return base.getBoolean(name_hide_mod, category_behavior, false, "Hide mod names on tooltips.");
+		return base.getBoolean("Hide Mod Name", category_behavior, false, "Hide mod names on tooltips. Enable if you see two mod names.");
 	}
 
 	public float getOpacity() {
-		return base.getFloat(name_opacity, category_appearance, 0.75F, 0.0F, 1.0F, "Set the opacity for the tooltips; 0 being completely invisible and 1 being completely opaque.");
+		return base.getFloat("Tooltip Opacity", category_appearance, 0.75F, 0.0F, 1.0F, "Set the opacity for the tooltips; 0 being completely invisible and 1 being completely opaque.");
 	}
 
 	public Property getOutline() {
-		return base.get(category_appearance, name_outline, "0x5000FF", "Choose a color using the gui by clicking the color button or type in a color manually.", Type.COLOR);
+		return base.get(category_appearance, "Outline Color", "0x5000FF", "Choose a color using the gui by clicking the color button or type in a color manually.", Type.COLOR);
 	}
 
 	public Property getBackground() {
-		return base.get(category_appearance, name_background, "0x100010", "Choose a color using the gui by clicking the color button or type in a color manually.", Type.COLOR);
+		return base.get(category_appearance, "Background Color", "0x100010", "Choose a color using the gui by clicking the color button or type in a color manually.", Type.COLOR);
 	}
 
 	public int getBackgroundColor() {
